nth catalan numbers.py
Here's a step-by-step explanation of the given Python code:

The function findCatalan takes an integer n as input and returns the nth Catalan number.

The code first checks if n is less than or equal to 1. If so, it returns 1. This is because the first two Catalan numbers are both 1.

The variable MOD is set to 10^9 + 7. This is used to take the result modulo this value to prevent overflow, as the numbers involved can become quite large.

An array catalan of size (n + 1) is created and initialized with zeros. This array will be used to store the calculated Catalan numbers.

The first two elements of catalan are set to 1 since the 0th and 1st Catalan numbers are both 1.

The loop runs from 2 to n (inclusive). This is because we have already set the values for the 0th and 1st Catalan numbers.

Within this loop, there's another nested loop that iterates through each value j from 0 to i - 1. This is to calculate the Catalan number for the current value of i.

The formula for calculating the nth Catalan number is:
**catalan[n] = sum(catalan[i] * catalan[n - i - 1]) for i in range(n)**

This formula considers all possible combinations of two smaller Catalan numbers to calculate the larger Catalan number.

The catalan[i] * catalan[n - i - 1] term is added to catalan[n] for each i. This corresponds to the summation part of the formula.

After calculating the nth Catalan number, the result is returned. It's important to use the modulo operation with MOD to ensure that the result doesn't exceed the specified value.